modeling_pipeline/pipeline_v5/requirements.txt
# V5 Pipeline Dependencies

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
ijson>=3.2.0

# Machine Learning
catboost>=1.2.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Database
psycopg2-binary>=2.9.0

# API & HTTP
requests>=2.31.0

# Environment
python-dotenv>=1.0.0

# Optional: Hyperparameter Tuning
optuna>=3.4.0

# Optional: Progress bars
tqdm>=4.66.0
